+++ numth/tests/polynomial_test.py
