+++ numth/tests/quadratic_test.py
